requirements.txt
awslambdaric
boto3
requests
python-dotenv